- Implemented isValidApiKeys guard
