+++ apps/charge-api-service/src/api-keys/api-keys.controller.ts
@@ -1,6 +1,7 @@
-import { Controller } from '@nestjs/common'
+import { Controller, Get, UseGuards } from '@nestjs/common'
 import { MessagePattern } from '@nestjs/microservices'
 import { ApiKeysService } from 'apps/charge-api-service/src/api-keys/api-keys.service'
+import { IsValidApiKeysGuard } from './guards/is-valid-api-keys.guard'
 
 @Controller()
 export class ApiKeysController {
@@ -54,5 +55,15 @@ export class ApiKeysController {
   @MessagePattern('get_public')
   getPublic (projectId: string) {
     return this.apiKeysService.getPublicKey(projectId)
+  }
+
+  /**
+   * TODO - Just for demonstrating the usage of IsValidApiKeysGuard.
+   * Remove this once we have actual APIs that use the Guard
+   */
+  @Get('/checkApiKeysValidity')
+  @UseGuards(IsValidApiKeysGuard)
+  checkApiKeys () {
+    return true
   }
 }
